move temperature,lighting and ph to modules folder
