+++ controller/modules/lighting/api_test.go
