+++ controller/modules/lighting/cycle_test.go
