+++ controller/modules/lighting/light.go
