+++ controller/modules/lighting/profile_test.go
