+++ controller/modules/temperature/sensor.go
