+++ controller/modules/temperature/sensor_test.go
