+++ controller/subsystem_loader.go
@@ -9,14 +9,14 @@ import (
 
 	"github.com/reef-pi/reef-pi/controller/connectors"
 	"github.com/reef-pi/reef-pi/controller/equipment"
-	"github.com/reef-pi/reef-pi/controller/lighting"
 	"github.com/reef-pi/reef-pi/controller/macro"
 	"github.com/reef-pi/reef-pi/controller/modules/ato"
 	"github.com/reef-pi/reef-pi/controller/modules/camera"
 	"github.com/reef-pi/reef-pi/controller/modules/doser"
-	"github.com/reef-pi/reef-pi/controller/ph"
+	"github.com/reef-pi/reef-pi/controller/modules/lighting"
+	"github.com/reef-pi/reef-pi/controller/modules/ph"
+	"github.com/reef-pi/reef-pi/controller/modules/temperature"
 	"github.com/reef-pi/reef-pi/controller/system"
-	"github.com/reef-pi/reef-pi/controller/temperature"
 	"github.com/reef-pi/reef-pi/controller/timer"
 )
